+++ R/vctrs.R
@@ -1,0 +1,12 @@
+# vctrs boilerplate
+#
+# Author: mjskay
+###############################################################################
+
+
+#' Internal vctrs methods
+#'
+#' @import vctrs
+#' @keywords internal
+#' @name ggdist-vctrs
+NULL